Bug 1912695: Clean up uses of offset() and offset64(). r=yury,rhunt
This patch attempts to clean up and clarify code surrounding offsets in
baseline and ion. No behavior was changed, except in the case of
removing some unnecessary conversions.

Differential Revision: https://phabricator.services.mozilla.com/D219931

UltraBlame original commit: 622c829c663473ad99c905c87c8026d3a64738d7

+++ js/src/wasm/WasmIonCompile.cpp
@@ -1361,7 +1361,7 @@ class FunctionCompiler {
   
   
   void foldConstantPointer(MemoryAccessDesc* access, MDefinition** base) {
-    uint32_t offsetGuardLimit = GetMaxOffsetGuardLimit(
+    uint64_t offsetGuardLimit = GetMaxOffsetGuardLimit(
         codeMeta_.hugeMemoryEnabled(access->memoryIndex()));
 
     if ((*base)->isConstant()) {
@@ -1373,9 +1373,8 @@ class FunctionCompiler {
       }
 
       uint64_t offset = access->offset64();
-
       if (offset < offsetGuardLimit && basePtr < offsetGuardLimit - offset) {
-        offset += uint32_t(basePtr);
+        offset += basePtr;
         access->setOffset32(uint32_t(offset));
         *base = isMem64(access->memoryIndex()) ? constantI64(int64_t(0))
                                                : constantI32(0);
@@ -1387,7 +1386,7 @@ class FunctionCompiler {
   
   void maybeComputeEffectiveAddress(MemoryAccessDesc* access,
                                     MDefinition** base, bool mustAddOffset) {
-    uint32_t offsetGuardLimit = GetMaxOffsetGuardLimit(
+    uint64_t offsetGuardLimit = GetMaxOffsetGuardLimit(
         codeMeta_.hugeMemoryEnabled(access->memoryIndex()));
 
     if (access->offset64() >= offsetGuardLimit ||
@@ -1541,6 +1540,7 @@ class FunctionCompiler {
     return codeMeta_.hugeMemoryEnabled(memoryIndex);
   }
 
+  
   
   MDefinition* computeEffectiveAddress(MDefinition* base,
                                        MemoryAccessDesc* access) {
